+++ hutool-captcha/src/test/java/cn/hutool/captcha/CaptchaTest.java
@@ -13,7 +13,7 @@ import cn.hutool.core.lang.Console;
 public class CaptchaTest {
 	
 	@Test
-//	@Ignore
+	@Ignore
 	public void lineCaptchaTest() {
 		
 		//定义图形验证码的长和宽