+++ hutool-core/src/main/java/cn/hutool/core/util/StrUtil.java
@@ -634,6 +634,31 @@ public class StrUtil {
 	public static boolean containsAny(CharSequence str, CharSequence... testStrs) {
 		return null != getContainsStr(str, testStrs);
 	}
+	
+	/**
+	 * 给定字符串是否包含空白符（空白符包括空格、制表符、全角空格和不间断空格）<br>
+	 * 如果给定字符串为null或者""，则返回false
+	 * 
+	 * @param str 字符串
+	 * @return 是否包含空白符
+	 * @since 4.0.8
+	 */
+	public static boolean containsBlank(CharSequence str) {
+		if(null == str) {
+			return false;
+		}
+		final int length = str.length();
+		if(0 == length) {
+			return false;
+		}
+		
+		for (int i = 0; i < length; i += 1) {
+			if (NumberUtil.isBlankChar(str.charAt(i))) {
+				return true;
+			}
+		}
+		return false;
+	}
 
 	/**
 	 * 查找指定字符串是否包含指定字符串列表中的任意一个字符串，如果包含返回找到的第一个字符串