+++ hutool-db/src/main/java/cn/hutool/db/DbUtil.java
@@ -606,7 +606,20 @@ public final class DbUtil {
 	}
 
 	/**
-	 * 创建LIKE语句中的值
+	 * 创建LIKE语句中的值，创建的结果为：
+	 * 
+	 * <pre>
+	 * 1、LikeType.StartWith: %value
+	 * 2、LikeType.EndWith: value%
+	 * 3、LikeType.Contains: %value%
+	 * </pre>
+	 * 
+	 * 如果withLikeKeyword为true，则结果为：
+	 * <pre>
+	 * 1、LikeType.StartWith: LIKE %value
+	 * 2、LikeType.EndWith: LIKE value%
+	 * 3、LikeType.Contains: LIKE %value%
+	 * </pre>
 	 * 
 	 * @param value 被查找值
 	 * @param likeType LIKE值类型 {@link LikeType}