+++ hutool-json/src/main/java/cn/hutool/json/JSONException.java
@@ -1,5 +1,6 @@
 package cn.hutool.json;
 
+import cn.hutool.core.exceptions.ExceptionUtil;
 import cn.hutool.core.util.StrUtil;
 
 /**
@@ -10,17 +11,21 @@ import cn.hutool.core.util.StrUtil;
  */
 public class JSONException extends RuntimeException {
 	private static final long serialVersionUID = 0;
+	
+	public JSONException(Throwable e) {
+		super(ExceptionUtil.getMessage(e), e);
+	}
 
-	public JSONException(final String message) {
+	public JSONException(String message) {
 		super(message);
 	}
-
-	public JSONException(final String message, final Throwable cause) {
-		super(message, cause);
+	
+	public JSONException(String messageTemplate, Object... params) {
+		super(StrUtil.format(messageTemplate, params));
 	}
 
-	public JSONException(final Throwable cause) {
-		super(cause.getMessage(), cause);
+	public JSONException(String message, Throwable cause) {
+		super(message, cause);
 	}
 
 	public JSONException(Throwable throwable, String messageTemplate, Object... params) {