+++ hutool-json/src/main/java/cn/hutool/json/JSONNull.java
@@ -1,13 +1,16 @@
 package cn.hutool.json;
 
+import java.io.Serializable;
+
 /**
  * 用于定义<code>null</code>，与Javascript中null相对应<br>
  * Java中的<code>null</code>值在js中表示为undefined。
  * @author Looly
  *
  */
-public class JSONNull {
-	
+public class JSONNull implements Serializable{
+	private static final long serialVersionUID = 2633815155870764938L;
+
 	/**
 	 * <code>NULL</code> 对象用于减少歧义来表示Java 中的<code>null</code> <br>
 	 * <code>NULL.equals(null)</code> 返回 <code>true</code>. <br>