+++ hutool-poi/src/main/java/cn/hutool/poi/excel/ExcelWriter.java
@@ -48,7 +48,7 @@ public class ExcelWriter implements Closeable {
 	private Map<String, String> headerAlias;
 	/** 样式集，定义不同类型数据样式 */
 	private StyleSet styleSet;
-
+	
 	// -------------------------------------------------------------------------- Constructor start
 	/**
 	 * 构造，默认生成xls格式的Excel文件<br>
@@ -255,6 +255,17 @@ public class ExcelWriter implements Closeable {
 		this.currentRow.set(0);
 		return this;
 	}
+	
+	/**
+	 * 切换sheet，如果指定的sheet不存在，创建之
+	 * @param sheetName sheet名
+	 * @return this
+	 * @since 4.0.8
+	 */
+	public ExcelWriter setOrCreateSheet(String sheetName) {
+		this.sheet = ExcelUtil.getOrCreateSheet(this.workbook, sheetName);
+		return this;
+	}
 
 	/**
 	 * 设置写出的目标文件
@@ -276,6 +287,40 @@ public class ExcelWriter implements Closeable {
 	 */
 	public ExcelWriter setHeaderAlias(Map<String, String> headerAlias) {
 		this.headerAlias = headerAlias;
+		return this;
+	}
+	
+	/**
+	 * 设置列宽（单位为一个字符的宽度，例如传入width为10，表示10个字符的宽度）
+	 * 
+	 * @param columnIndex 列号（从0开始计数，-1表示所有列的默认宽度）
+	 * @param width 宽度（单位1~256个字符宽度）
+	 * @return this
+	 * @since 4.0.8
+	 */
+	public ExcelWriter setColumnWidth(int columnIndex, int width) {
+		if(columnIndex < 0) {
+			this.sheet.setDefaultColumnWidth(width);
+		}else {
+			this.sheet.setColumnWidth(columnIndex, width * 256);
+		}
+		return this;
+	}
+	
+	/**
+	 * 设置行高，值为一个点的高度
+	 * 
+	 * @param rownum 行号（从0开始计数，-1表示所有行的默认高度）
+	 * @param height 高度
+	 * @return this
+	 * @since 4.0.8
+	 */
+	public ExcelWriter setRowHeight(int rownum, int height) {
+		if(rownum < 0) {
+			this.sheet.setDefaultRowHeightInPoints(height);
+		}else {
+			this.sheet.getRow(rownum).setHeightInPoints(height);
+		}
 		return this;
 	}