+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 -------------------------------------------------------------------------------------------------------------
 
+## 4.0.8
+
+### 新特性
+* 【core】新增PinyinComparator、CollUtil新增sortByPinyin（感谢@【帝都】宁静）
+* 【json】JSONUtil增加xmlToJson方法
+* 【poi】 ExcelWriter增加setColumnWidth和setRowHeight方法
+* 【core】FileUtil.clean增加字符串重载（感谢@【帝都】宁静）
+* 【core】ArrayUtil增加insert方法（感谢@【帝都】宁静）
+* 【core】RandomUtil.randomDouble增加可选保留小数重载（感谢@【帝都】宁静）
+* 【core】增加RandomUtil.randomDay随机天（感谢@【帝都】宁静）
+* 【poi】  ExcelWriter增加setOrCreateSheet方法，从而支持多sheet生成
+
+### Bug修复
+* 【json】修复JSONArray中addAll加入两次的bug（感谢@【天津】〓下页）
+* 【core】修复BeanDesc中对static属性未忽略的问题（感谢@【深圳】枫林晓寒）
+* 【http】解决无法移除默认头信息的问题
+* 【core】修复Base64在decode时针对urlSafe乱码问题（issue#89@Github）
+* 【core】修复ReUtil.extractMulti(感谢@【杭州】徐承恩)
+* 【core】修复DESede类中算法错误问题（issue#93@Github）
+
+-------------------------------------------------------------------------------------------------------------
+
 ## 4.0.7
 
 ### 新特性